+++ src/normalizer.ts
@@ -5,13 +5,16 @@ import type * as ast from "./ast.ts"
 
 /**
  * Normalize `schema`.
- * This function should be called before code generation.
+ *
+ * It assigns aliases to certain compound types in an attempt to dedup them.
+ *
+ * This function should be called before JS/TS code generation.
  */
 export function normalize(schema: ast.Ast): ast.Ast {
     const n: Context = { defs: [], dedup: new Map(), aliasCount: 0 }
     const defs = n.defs
     for (const def of schema.defs) {
-        const type = normalizeSubtypes(n, def.type)
+        const type = normalizeType(n, def.type)
         if (def.type !== type) {
             const { alias, internal, comment, offset } = def
             defs.push({ alias, internal, type, comment, offset })
@@ -25,12 +28,27 @@ export function normalize(schema: ast.Ast): ast.Ast {
 }
 
 type Context = {
+    /**
+     * Aliiased types.
+     */
     readonly defs: ast.AliasedType[]
-    readonly dedup: Map<unknown, string>
+    /**
+     * Map a stringified type to its alias.
+     * The types are stringified using `JSON.stringify`.
+     * Offsets are set to `0`.
+     *
+     * This enables to deduiplicate types,
+     * so two identical types are assigned to the same alias.
+     */
+    readonly dedup: Map<string, string>
+    /**
+     * Number of generated aliases.
+     * Tracking their number allows us to generate unique aliases.
+     */
     aliasCount: number
 }
 
-function normalizeSubtypes(n: Context, type: ast.Type): ast.Type {
+function normalizeType(n: Context, type: ast.Type): ast.Type {
     if (type.types != null && type.types.length > 0) {
         const types = type.types.map((t) => maybeAlias(n, t))
         if (type.types.some((t, i) => t !== types[i])) {
@@ -43,6 +61,9 @@ function normalizeSubtypes(n: Context, type: ast.Type): ast.Type {
     return type
 }
 
+/**
+ * Normalize subtypes and alias lists, maps, optionals and unions.
+ */
 function maybeAlias(n: Context, type: ast.Type): ast.Type {
     switch (type.tag) {
         case "list": {
@@ -56,9 +77,14 @@ function maybeAlias(n: Context, type: ast.Type): ast.Type {
         case "union":
             return genAlias(n, type)
     }
-    return normalizeSubtypes(n, type)
+    return normalizeType(n, type)
 }
 
+/**
+ * Generate a new alias for `type` if the type doesn't exist.
+ * Otherwise, use the alias that has already been assigned to
+ * a type with the same shape as `type`.
+ */
 function genAlias(n: Context, type: ast.Type): ast.Alias {
     // NOTE: this dirty check is ok because we initialize
     // every object in the same way (properties are in the same order)
@@ -68,7 +94,7 @@ function genAlias(n: Context, type: ast.Type): ast.Alias {
     )
     let alias = n.dedup.get(stringifiedType)
     if (alias == null) {
-        const normalized = normalizeSubtypes(n, type)
+        const normalized = normalizeType(n, type)
         // We use an integer as internal alias.
         // This avoids conflicts with user aliases.
         alias = `${n.aliasCount++}`